refactor: CurrentUserId 어노테이션 패키지 위치 이동
