+++ src/main/java/com/campus/campus/domain/user/presentation/AuthController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.campus.campus.global.auth.annotation.CurrentUserId;
+import com.campus.campus.global.annotation.CurrentUserId;
 import com.campus.campus.global.auth.application.dto.OauthLoginResponse;
 import com.campus.campus.domain.user.application.service.KakaoOauthService;
 import com.campus.campus.global.common.response.CommonResponse;
